% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{as.character.css_property}
\alias{as.character.css_property}
\title{Convert a CSS 'property' object to a string}
\usage{
\method{as.character}{css_property}(x, ...)
}
\arguments{
\item{x}{property object}

\item{...}{other arguments}
}
\description{
Convert a CSS 'property' object to a string
}
