% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selector.R
\name{as.character.Selector}
\alias{as.character.Selector}
\title{Retrieve the character representation of a Selector object}
\usage{
\method{as.character}{Selector}(x, ...)
}
\arguments{
\item{x}{Selector object}

\item{...}{other arguments}
}
\description{
Retrieve the character representation of a Selector object
}
