% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\name{as.character.Style}
\alias{as.character.Style}
\title{Retrieve the character representation of a Style object}
\usage{
\method{as.character}{Style}(x, ...)
}
\arguments{
\item{x}{Style object}

\item{...}{other arguments}
}
\description{
Retrieve the character representation of a Style object
}
