% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleSheet.R
\name{as.character.StyleSheet}
\alias{as.character.StyleSheet}
\title{Retrieve the character representation of a SStyleSheettyle object}
\usage{
\method{as.character}{StyleSheet}(x, ...)
}
\arguments{
\item{x}{StyleSheet object}

\item{...}{other arguments}
}
\description{
Retrieve the character representation of a SStyleSheettyle object
}
