% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\docType{data}
\name{css_prop}
\alias{css_prop}
\title{CSS property helper}
\format{An object of class \code{list} of length 238.}
\usage{
css_prop
}
\description{
Uses autocomplete to help write some standard propertys
}
\keyword{datasets}
