% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selector.R
\name{css_sel}
\alias{css_sel}
\title{Helper for creating a Selector object}
\usage{
css_sel(selector = NULL)
}
\arguments{
\item{selector}{string or another Selector object to clone}
}
\description{
Helper for creating a Selector object
}
