% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\name{css_style}
\alias{css_style}
\title{Helper function for creating a Style object for a CSS style (requires selector)}
\usage{
css_style(..., selector = NULL)
}
\arguments{
\item{...}{named list of properties for the style e.g. `color = "black"`}

\item{selector}{character string or object of Selector class}
}
\description{
Helper function for creating a Style object for a CSS style (requires selector)
}
