% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{imp}
\alias{imp}
\title{Helper to manually create new "!important" properies}
\usage{
imp(...)
}
\arguments{
\item{...}{a name/value pair}
}
\description{
Helper to manually create new "!important" properies
}
