% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyframes.R
\docType{data}
\name{Keyframe}
\alias{Keyframe}
\alias{css_keyframe}
\title{Create a single keyframe for a style at a moment in time.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Keyframe
}
\description{
This is identical to the \code{Style} class except the first argument is a
time specification (either 'from', 'to', or a percentage), instead of a name
}
\examples{
\dontrun{
Keyframe$new('from', color = '#000000')
Keyframe$new('to'  , color = '#ffffff')
}

}
\keyword{datasets}
