% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyframes.R
\docType{data}
\name{Keyframes}
\alias{Keyframes}
\alias{css_keyframes}
\title{Create a named Keyframes style suitable to insert in a StyleSheet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Keyframes
}
\description{
Create a named Keyframes style suitable to insert in a StyleSheet
}
\keyword{datasets}
