% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{print.css_property}
\alias{print.css_property}
\title{Print a CSS 'property' object}
\usage{
\method{print}{css_property}(x, ...)
}
\arguments{
\item{x}{property object}

\item{...}{other arguments}
}
\description{
Print a CSS 'property' object
}
