% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selector.R
\docType{class}
\name{Selector}
\alias{Selector}
\title{CSS selector builder}
\description{
Based upon a naive reading of \url{https://developer.mozilla.org/en-US/docs/Web/CSS/Reference#Selectors}.
}
\section{Usage}{

\preformatted{selector <- Selector$new()
selector <- css_sel()

selector$type(type_name)
selector$class(...)
selector$id(id_name)

new_selector <- selector$copy()

}
}

\section{Methods}{


\describe{

\item{\code{$type(type_name)}}{
Set the type for the selector
\tabular{ll}{
  \code{type_name} \tab type name \cr
}
}

\item{\code{$id(id_name)}}{
Set the id for the selector
\tabular{ll}{
  \code{id_name} \tab id name \cr
}
}

\item{\code{$class(...)}}{
Add to the class list for for the selector
\tabular{ll}{
  \code{...} \tab class names \cr
}
}

\item{\code{$attr(..., modifier = c('exact', 'choice', 'prefix', 'suffix', 'contains'))}}{
Add attribute matching to the selector. See \url{https://developer.mozilla.org/en-US/docs/Web/CSS/Attribute_selectors}
\tabular{ll}{
  \code{...} \tab name/value pairs for the atributes \cr
  \code{modifer} \tab by default 'exact' attribute matching, but can select from among the different types of matching  \cr
}
}

\item{\code{$pseudo_class(...)}}{
Set pseudo classes on the selector.
\tabular{ll}{
  \code{...} \tab Use \code{name = TRUE} to set pseudo-classes \cr
}
}

\item{\code{$pseudo_element(...)}}{
Set pseudo element on the selector.
\tabular{ll}{
  \code{...} \tab Use \code{name = TRUE} to set pseudo-elements \cr
}
}



\item{\code{$descendant_of(selector), $parent_of(selector), $child_of(selector), $prior_adjacent(selector), $prior_sibling(selector), $and(selector)}}{
Set relationship with another selector
\tabular{ll}{
  \code{selector} \tab Other selector \cr
}
}

\item{\code{$as_character()}}{
Convert \code{Selector} to a character string.
}

\item{\code{$print()}}{
Print \code{Selector} to terminal.
}

\item{\code{$copy()}}{
Copy \code{Selector}.
}

}
}

\examples{
\dontrun{
selector <- css_sel(".marsha")$child_of('carol')
}

}
\keyword{datasets}
