% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleSheet.R
\docType{class}
\name{StyleSheet}
\alias{StyleSheet}
\alias{css_stylesheet}
\title{CSS style sheet builder}
\description{
CSS style sheet builder
}
\section{Usage}{

\preformatted{sheet <- StyleSheet$new()
sheet <- css_stylesheet()

sheet$append(style)
sheet$remove(3)
sheet$add('h1', color = 'blue')
sheet$as_character()
new_sheet <- sheet$copy()

}
}

\section{Methods}{


\describe{

\item{\code{$new(...)}}{
Initialise a stylesheet
\tabular{ll}{
  \code{...} \tab all arguments treated as \code{Style} objects and added to the style sheet \cr
}
}


\item{\code{$append(...)}}{
Append \code{Style} objects to this StyleSheet
\tabular{ll}{
  \code{...} \tab all arguments treated as Style objects and added to the style sheet \cr
}
}

\item{\code{$remove(indices)}}{
Remove \code{Style} objects from this StyleSheet by index.
\tabular{ll}{
  \code{indices} \tab indices of Style objects to remove \cr
}
}

\item{\code{$add(..., selector = NULL)}}{
Creates a \code{Style} object and adds it to the StyleSheet.  Returns the newly
created \code{Style}.
\tabular{ll}{
  \code{...} \tab name/value properties \cr
  \code{selector} \tab If given, then used as the the selector for this style. If
  not given, then the first unnamed argument in \code{...} is interpreted as the
  character string for the selector. \cr
}
}

\item{\code{$as_character()}}{
Convert \code{StyleSheet} to a character string.
}

\item{\code{$print()}}{
Print \code{StyleSheet} to terminal.
}

\item{\code{$copy()}}{
Copy \code{StyleSheet}.
}

}
}

\examples{
\dontrun{
marsha <- css_style('#marsha')$
  update(marsha = 'marsha')$
  update(ptag$display$none, width = "100\%")

sheet <- StyleSheet$new(marsha)
}

}
\keyword{datasets}
